Restructure packages in the Logic component
Rearranged the file structure to avoid packages having names longer than a single word.
