+++ src/main/java/seedu/address/logic/PshLogicManager.java
@@ -7,11 +7,11 @@ import java.util.logging.Logger;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.generalcommands.Command;
-import seedu.address.logic.generalcommands.CommandResult;
-import seedu.address.logic.generalcommands.exceptions.CommandException;
-import seedu.address.logic.generalparser.exceptions.ParseException;
-import seedu.address.logic.petparser.PetTrackerParser;
+import seedu.address.logic.commands.general.Command;
+import seedu.address.logic.commands.general.CommandResult;
+import seedu.address.logic.commands.general.exceptions.CommandException;
+import seedu.address.logic.parser.general.exceptions.ParseException;
+import seedu.address.logic.parser.pet.PetTrackerParser;
 import seedu.address.model.PshModel;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyPetTracker;
