+++ src/main/java/seedu/address/logic/commands/general/ExitCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.generalcommands;
+package seedu.address.logic.commands.general;
 
 import seedu.address.model.PshModel;
