+++ src/main/java/seedu/address/logic/commands/pet/EditPetCommand.java
@@ -1,12 +1,12 @@
-package seedu.address.logic.petcommands;
+package seedu.address.logic.commands.pet;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.petparser.CliSyntax.PREFIX_DOB;
-import static seedu.address.logic.petparser.CliSyntax.PREFIX_FOODLIST;
-import static seedu.address.logic.petparser.CliSyntax.PREFIX_GENDER;
-import static seedu.address.logic.petparser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.petparser.CliSyntax.PREFIX_SPECIES;
-import static seedu.address.logic.petparser.CliSyntax.PREFIX_TAG;
+import static seedu.address.logic.parser.pet.CliSyntax.PREFIX_DOB;
+import static seedu.address.logic.parser.pet.CliSyntax.PREFIX_FOODLIST;
+import static seedu.address.logic.parser.pet.CliSyntax.PREFIX_GENDER;
+import static seedu.address.logic.parser.pet.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.parser.pet.CliSyntax.PREFIX_SPECIES;
+import static seedu.address.logic.parser.pet.CliSyntax.PREFIX_TAG;
 import static seedu.address.model.PshModel.PREDICATE_SHOW_ALL_PETS;
 
 import java.util.Collections;
@@ -18,9 +18,9 @@ import java.util.Set;
 import seedu.address.commons.core.PshMessages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
-import seedu.address.logic.generalcommands.Command;
-import seedu.address.logic.generalcommands.CommandResult;
-import seedu.address.logic.generalcommands.exceptions.CommandException;
+import seedu.address.logic.commands.general.Command;
+import seedu.address.logic.commands.general.CommandResult;
+import seedu.address.logic.commands.general.exceptions.CommandException;
 import seedu.address.model.PshModel;
 import seedu.address.model.pet.DateOfBirth;
 import seedu.address.model.pet.Food;
