+++ src/main/java/seedu/address/logic/commands/pet/FindPetCommand.java
@@ -1,11 +1,11 @@
-package seedu.address.logic.petcommands;
+package seedu.address.logic.commands.pet;
 
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.PshMessages;
-import seedu.address.logic.generalcommands.Command;
-import seedu.address.logic.generalcommands.CommandResult;
-import seedu.address.logic.generalcommands.exceptions.CommandException;
+import seedu.address.logic.commands.general.Command;
+import seedu.address.logic.commands.general.CommandResult;
+import seedu.address.logic.commands.general.exceptions.CommandException;
 import seedu.address.model.PshModel;
 import seedu.address.model.pet.NameContainsKeywordsPredicate;
